KD-Tree for Objects
KD-Tree representation of top level objects in the scene. Should be
helpful for scenes with many objects, although my test scene didn't show
any noticeable performance boost (only 9 objects). Took a while to get
the tree working - not sure if this initial drop is 100% working, but
close enough that the output image looks the same as with the old
algorithm.